pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-api"
version = "25.13.8"
description = "GitLab API Python Wrapper"
readme = "README.md"
authors = [{ name = "Audel Rouhi", email = "knucklessg1@gmail.com" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: Public Domain",
    "Environment :: Console",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.8.1",
    "urllib3>=2.2.2",
    "pydantic[email]>=2.8.2",
    "requests-toolbelt>=1.0.0",
]
[project.optional-dependencies]
mcp = [
    "fastmcp>=2.13.0.2",
    "eunomia-mcp>=0.3.10",
    "gql>=4.0.0",
]
a2a = [
    "pydantic-ai-slim[fastmcp,openai,anthropic,google,huggingface,a2a]>=1.32.0",
    "pydantic-ai-skills",
]
gql = [
    "gql>=4.0.0",
]
all = [
    "fastmcp>=2.13.0.2",
    "eunomia-mcp>=0.3.10",
    "pydantic-ai-slim[fastmcp,openai,anthropic,google,huggingface,a2a]>=1.32.0",
    "pydantic-ai-skills",
    "gql>=4.0.0",
]

[project.scripts]
gitlab-mcp = "gitlab_api.gitlab_mcp:gitlab_mcp"
gitlab-a2a = "gitlab_api.gitlab_a2a:agent_server"

[tool.setuptools.package-data]
"gitlab_api" = ["skills/**"]

[tool.setuptools.packages.find]
where = ["."]
